wip: type alias in types.rs
